requirements.txt
pandas>=0.25.3
urllib3>=1.25.8
request>==2.23.0
seaborn>=0.10.0
tabula_py>=2.1.0
matplotlib>=3.1.2
numpy>=1.18.2
scikit_learn>=0.22.2.post1
scipy>=1.4.1
tabula>=1.0.5
